requirements.txt
httpx>=0.27.0
PyMuPDF>=1.24.0
Pillow>=10.0.0
pyyaml>=6.0.0
olmocr[bench]>=0.4.0
litellm>=1.50.0
